Sections: shared component. Scroll implementd
