+++ app/layout/scenarios/edit/analysis/component.tsx
@@ -15,12 +15,29 @@ import Pill from 'layout/pill';
 import AdjustPanningUnits from 'layout/scenarios/edit/analysis/adjust-planning-units';
 import CostSurface from 'layout/scenarios/edit/analysis/cost-surface';
 import GapAnalysis from 'layout/scenarios/edit/analysis/gap-analysis';
-import Sections from 'layout/scenarios/edit/analysis/sections';
 import Run from 'layout/scenarios/edit/run';
+import Sections from 'layout/sections';
 
 import Button from 'components/button';
 import Modal from 'components/modal';
 
+const SECTIONS = [
+  {
+    id: 'gap-analysis',
+    name: 'Gap analysis',
+    description: 'A gap analysis calculates how much of each feature is under the existing protected area network and then summarises the representation as a percentage',
+  },
+  {
+    id: 'cost-surface',
+    name: 'Cost surface',
+    description: 'Costs reflect any variety of socioeconomic factors, which if minimized, might help the conservation plan be implemented more effectively and reduce conflicts with other uses.',
+  },
+  {
+    id: 'adjust-planning-units',
+    name: 'Adjust planning units (optional)',
+    description: 'The status of a planning unit determines whether it is included in every solution (i.e. locked in) or excluded (i.e. locked out). The default status is neither included or excluded but determined during the Marxan analysis.',
+  },
+];
 export interface ScenariosSidebarEditAnalysisProps {
 
 }
@@ -115,6 +132,7 @@ export const ScenariosSidebarEditAnalysis: React.FC<ScenariosSidebarEditAnalysis
               {!section && (
               <Sections
                 key="sections"
+                sections={SECTIONS}
                 onChangeSection={onChangeSection}
               />
               )}
