--- app/layout/scenarios/edit/analysis/sections/item/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
