+++ app/layout/scenarios/show/analysis/component.tsx
@@ -15,7 +15,25 @@ import Pill from 'layout/pill';
 import AdjustPanningUnitsShow from 'layout/scenarios/show/analysis/adjust-planning-units';
 import CostSurface from 'layout/scenarios/show/analysis/cost-surface';
 import GapAnalysis from 'layout/scenarios/show/analysis/gap-analysis';
-import Sections from 'layout/scenarios/show/analysis/sections';
+import Sections from 'layout/sections';
+
+const SECTIONS = [
+  {
+    id: 'gap-analysis',
+    name: 'Gap analysis',
+    description: 'A gap analysis calculates how much of each feature is under the existing protected area network and then summarises the representation as a percentage',
+  },
+  {
+    id: 'cost-surface',
+    name: 'Cost surface',
+    description: 'Costs reflect any variety of socioeconomic factors, which if minimized, might help the conservation plan be implemented more effectively and reduce conflicts with other uses.',
+  },
+  {
+    id: 'adjust-planning-units',
+    name: 'Adjust planning units (optional)',
+    description: 'The status of a planning unit determines whether it is included in every solution (i.e. locked in) or excluded (i.e. locked out). The default status is neither included or excluded but determined during the Marxan analysis.',
+  },
+];
 
 export interface ScenariosSidebarShowAnalysisProps {
 }
@@ -107,10 +125,11 @@ export const ScenariosSidebarShowAnalysis: React.FC<ScenariosSidebarShowAnalysis
               </header>
 
               {!section && (
-              <Sections
-                key="sections"
-                onChangeSection={onChangeSection}
-              />
+                <Sections
+                  key="sections"
+                  sections={SECTIONS}
+                  onChangeSection={onChangeSection}
+                />
               )}
 
               {section === 'gap-analysis' && (
