--- app/layout/scenarios/show/analysis/sections/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
