--- app/layout/scenarios/show/analysis/sections/item/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
