--- app/layout/scenarios/show/solutions/sections/constants.ts
@@ -1,20 +1,0 @@
-import { SolutionsSections } from './types';
-
-export const SECTIONS = [
-  {
-    id: SolutionsSections.DETAILS,
-    name: 'Details',
-    description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit quis quisquam, reiciendis neque, facere perspiciatis.',
-  },
-  {
-    id: SolutionsSections.GAP_ANALYSIS,
-    name: 'Gap Analysis',
-    description: 'Lorem ipsum dolor sit amet, consectetur adipisicing elit. Fugit quis quisquam, reiciendis neque, facere perspiciatis.',
-  },
-  // {
-  //   id: SolutionsSections.SOLUTIONS,
-  //   name: 'Solutions',
-  //   description: 'Lorem ipsum dolor sit amet, consectetur
-  //   adipisicing elit. Fugit quis quisquam, reiciendis neque, facere perspiciatis.',
-  // },
-];
