--- app/layout/scenarios/show/solutions/sections/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
