--- app/layout/scenarios/show/solutions/sections/item/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
