--- app/layout/scenarios/show/solutions/sections/types.ts
@@ -1,9 +1,0 @@
-export interface ScenariosSidebarSolutionsSectionsProps {
-  onChangeSection: (s: string) => void;
-}
-
-export enum SolutionsSections {
-  DETAILS = 'details',
-  GAP_ANALYSIS = 'gap-analysis',
-  SOLUTIONS = 'solutions',
-}
