+++ app/layout/sections/item/component.tsx
@@ -10,6 +10,7 @@ export interface ScenariosSidebarAnalysisSectionProps {
   id: string;
   name: string;
   description: string;
+  readonly?: boolean,
   onChangeSection: (s: string) => void;
 }
 
@@ -17,11 +18,14 @@ export const ScenariosSidebarAnalysisSection: React.FC<ScenariosSidebarAnalysisS
   id,
   name,
   description,
+  readonly,
   onChangeSection,
 }: ScenariosSidebarAnalysisSectionProps) => {
   const [animate, setAnimate] = useState('leave');
 
   return (
+    !readonly && (
+
     <button
       type="button"
       className="relative w-full py-5 pr-20 text-left focus:outline-none"
@@ -35,30 +39,37 @@ export const ScenariosSidebarAnalysisSection: React.FC<ScenariosSidebarAnalysisS
       <motion.div
         className="absolute transform -translate-y-1/2 right-4 top-1/2"
         initial={{
-          opacity: 0,
-          x: -10,
+          opacity: 1,
+          x: 0,
           y: '-50%',
         }}
         animate={animate}
         variants={{
           enter: {
             opacity: 1,
-            x: 0,
+            x: 5,
             y: '-50%',
+            transition: {
+              ease: 'easeInOut',
+              repeat: Infinity,
+              repeatType: 'mirror',
+              duration: 0.4,
+            },
           },
           leave: {
-            opacity: 0,
-            x: -10,
+            opacity: 1,
+            x: 0,
             y: '-50%',
+            transition: {
+              ease: 'easeInOut',
+            },
           },
-        }}
-        transition={{
-          ease: 'anticipate',
         }}
       >
         <Icon icon={ARROW_RIGHT_2_SVG} className="w-5 h-5 opacity-75 text-primary-500" />
       </motion.div>
     </button>
+    )
   );
 };
